requirements.txt
sentinelhub==3.11.1
shapely==2.1.0
geopandas==1.1.1
numpy==2.3.0
pandas==3.0.1
langchain==1.2.0
streamlit==1.54.0
python-dotenv==1.2.2
requests==2.32.5
rasterio==1.4.3
openai
langchain
pinecone-client
streamlit


